adapt unit-test for dropping base64

+++ solidity/bmv/contracts/libraries/RLPEncode.sol
@@ -72,7 +72,7 @@ library RLPEncode {
      * @return The RLP encoded uint in bytes.
      */
     function encodeUint(uint256 self) internal pure returns (bytes memory) {
-        uint nBytes = bitLength(self)/8 + 1;
+        uint256 nBytes = bitLength(self) / 8 + 1;
         bytes memory uintBytes = encodeUintByLength(self);
         if (nBytes - uintBytes.length > 0) {
             uintBytes = abi.encodePacked(bytes1(0), uintBytes);
@@ -351,7 +351,10 @@ library RLPEncode {
         } else if (length >= MAX_UINT112 && length < MAX_UINT120) {
             return abi.encodePacked(uint120(length));
         }
-        require(length >= MAX_UINT120 && length < MAX_UINT128, "outOfBounds: [0, 2^128]");
+        require(
+            length >= MAX_UINT120 && length < MAX_UINT128,
+            "outOfBounds: [0, 2^128]"
+        );
         return abi.encodePacked(uint128(length));
     }
